+++ src/components/switch/slider.ts
@@ -149,6 +149,7 @@ export class CuiSliderHandler extends CuiMutableHandler<CuiSliderArgs> implement
                 this.#startX = data.x;
                 this.#currSlider.setElement(current);
                 this.helper.setClassesAs(document.body, CLASSES.swipingOn);
+                data.event.preventDefault();
                 break;
             case "up":
                 if (!this.#isTracking) {
@@ -191,6 +192,7 @@ export class CuiSliderHandler extends CuiMutableHandler<CuiSliderArgs> implement
                     this.#currSlider.update(Math.abs(this.#swipeRatio));
                     this.#nextSlider.update(Math.abs(this.#swipeRatio));
                 })
+                data.event.preventDefault();
                 break;
             default:
                 break;