+++ src/core/models/colors.ts
@@ -1,4 +1,4 @@
-import { getRangeValue, increaseValue, decreaseValue } from "../utils/functions";
+import { getRangeValue, } from "../utils/functions";
 
 export class CuiColor {
     #red: number;
@@ -15,21 +15,38 @@ export class CuiColor {
         this.#red = getRangeValue(red, 0, 255);
         this.#green = getRangeValue(green, 0, 255);
     }
+
+    setRed(red: number) {
+        this.#red = getRangeValue(red, 0, 255);
+    }
+
+    setGreen(green: number) {
+        this.#green = getRangeValue(green, 0, 255);
+    }
+
+    setBlue(blue: number) {
+        this.#blue = getRangeValue(blue, 0, 255);
+    }
+
+    opacity(val: number): CuiColor {
+        this.#alpha = getRangeValue(val, 0, 1);
+        return this;
+    }
+
     lighten(amount: number): CuiColor {
-        let percent = getRangeValue(amount, 0, 100) / 100;
-        this.set(increaseValue(this.#red, percent),
-            increaseValue(this.#green, percent),
-            increaseValue(this.#blue, percent),
-            this.#alpha)
+        this.shade(amount)
         return this;
     }
 
     darken(amount: number): CuiColor {
-        let percent = getRangeValue(amount, 0, 100) / 100;
-        this.set(decreaseValue(this.#red, percent),
-            decreaseValue(this.#green, percent),
-            decreaseValue(this.#blue, percent),
-            this.#alpha)
+        this.shade(-amount);
+        return this;
+    }
+
+    invert(): CuiColor {
+        this.#blue = 255 - this.#blue;
+        this.#red = 255 - this.#red;
+        this.#green = 255 - this.#green;
         return this;
     }
 
@@ -49,7 +66,20 @@ export class CuiColor {
     }
 
     toCssString(): string {
-        return `rgba(${this.#red},${this.#green},${this.#blue}, ${this.#alpha})`;
+        return `rgba(${this.#red}, ${this.#green}, ${this.#blue}, ${this.#alpha})`;
+    }
+
+    private shade(percent: number, self: boolean = true) {
+        this.#red = this.shadeSingle(this.#red, percent, self);
+        this.#green = this.shadeSingle(this.#green, percent, self);
+        this.#blue = this.shadeSingle(this.#blue, percent, self);
+    }
+
+    private shadeSingle(val: number, percent: number, self: boolean = true) {
+        let rel = self ? val : 255;
+        let prop = (rel * percent) / 100
+        let newVal = val + Math.round(prop);
+        return getRangeValue(newVal, 0, 255);
     }
 
     clone() {