+++ src/core/models/utils.ts
@@ -3,14 +3,13 @@ import { CuiSetup, CuiSetupInit } from "./setup";
 import { CuiLightMode } from "../utils/types";
 import { CuiInteractionsFactory } from "../factories/interactions";
 import { CuiCacheManager } from "../../app/managers/cache";
-import { CuiEventBus, CuiEventBusFactory } from "../bus/bus";
-import { TaskedEventEmitHandler } from "../bus/handlers";
-import { CuiCallbackExecutor } from "../bus/executors";
-import { getName, replacePrefix } from "../utils/functions";
+import { CuiEventBusFactory } from "../bus/bus";
+import { are, getName, replacePrefix } from "../utils/functions";
 import { CLASSES } from "../utils/statics";
 import { ICuiDocumentStyleAppender, CuiDocumentStyleAppender } from "../styles/appender";
 import { CuiInstanceColorHandler } from "../../app/handlers/colors";
 import { ICuiResizableObserver, CuiResizeObserver } from "../../app/observers/resize";
+import { CSSVariableError } from "./errors";
 
 export class CuiUtils {
     interactions: IUIInteractionProvider;
@@ -68,6 +67,9 @@ export class CuiUtils {
     }
 
     setProperty(name: string, value: string) {
+        if (!are(name, value)) {
+            throw new CSSVariableError("Property or value was not provided");
+        }
         let prop = replacePrefix(name, this.setup.prefix);
         document.documentElement.style.setProperty(prop, value);
     }